parsing

+++ ft_substr.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char	*str;
-	size_t	t;
+	int	t;
 
 	if (s == 0)
 		return (0);
